Bug cross delete kontogang gefixt diverse zum zum debuggen genutze console.log´s entfernt

+++ src/components/kontogangLoeschen.vue
@@ -12,6 +12,9 @@
           <h5 class="modal-title" id="exampleModalLabel">Transaktion Löschen</h5>
           <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
         </div>
+        <div class="alert alert-danger" role="alert" v-if="showAlert">
+          Id ist nicht bekannt guck nochmal genau
+        </div>
         <div class="modal-body">
           <div class="form-floating">
             <input v-model="id" class="form-control"  placeholder="TransaktionsId">
@@ -35,22 +38,38 @@ export default {
 
   data(){
     return {
-      id : 0
+      id : 0,
+      gibts : false,
+      showAlert : false
     }
   },
 
   methods: {
     deleteTransaktion(){
-      axios.delete(`http://localhost:8080/api/user/zugang/${this.id}`).then(async function (){
-        const {data} = await axios.get('http://localhost:8080/api/user');
-        store.kontostandId = data
-        for (let i = 0; i < data.kontostandIDs.length; i++) {
-          await axios.get(`http://localhost:8080/api/user/zugang/${data.kontostandIDs[i]}`).then(function(response){
-            store.newAmount += response.data.amount;
-          })
+      this.gibts = false
+      this.showAlert = false
+
+      for (let i = 0; i < store.kontostandId.kontostandIDs.length; i++){
+        if(store.kontostandId.kontostandIDs[i]== this.id){
+          this.gibts = true
         }
-        store.table_reload++;
-      })
+      }
+      if (this.gibts){
+        axios.delete(`http://localhost:8080/api/user/zugang/${this.id}`).then(async function (){
+          const {data} = await axios.get('http://localhost:8080/api/user');
+          store.kontostandId = data
+          store.newAmount = 0
+          for (let i = 0; i < data.kontostandIDs.length; i++) {
+            await axios.get(`http://localhost:8080/api/user/zugang/${data.kontostandIDs[i]}`).then(function(response){
+              store.newAmount += response.data.amount;
+            })
+          }
+          store.table_reload++;
+        })
+      } else {
+        this.gibts = false
+        this.showAlert = true
+      }
     }
   }
